% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleExtensions.R
\name{dgammaStateValueMembership}
\alias{dgammaStateValueMembership}
\title{Probability Density of an Ecosystem-State Value Distribution (Gamma Error)}
\usage{
dgammaStateValueMembership(x, stateVal, statePrec, stateProb, log = 0)
}
\arguments{
\item{x}{A scalar value containing the value of the ecosystem state variable}

\item{stateVal}{A numeric vector containing the mean values of the ecosystem state variable
associated with each alternative stable state}

\item{statePrec}{A numeric vector containing the precision (reciprocal of the variance) of the
error of the ecosystem state variable associated with each alternative stable state}

\item{stateProb}{A numeric vector containing the probabilities of being in each ecosystem stable
state (internally normalised to one)}

\item{log}{If \code{TRUE} then return the log density}
}
\value{
A scalar containing the probability density
}
\description{
Calculate the probability density of an ecosystem-state value distribution with a
gamma error distribution for the ecosystem state variable.
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
