% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmNIMBLE.R
\name{glmNIMBLE}
\alias{glmNIMBLE}
\title{Run Gernalised Linear Model in NIMBLE}
\usage{
glmNIMBLE(
  modelFormula,
  inputData,
  errorFamily = gaussian,
  regCoeffs = "none",
  modelSuffix = "",
  mcmcParams = list()
)
}
\arguments{
\item{modelFormula}{A \code{formula} project defining the relationship between the
response variable and the predictor variables in a similar manner to that used in
the \code{\link[base::glm]{glm}} function}

\item{inputData}{A \code{data.frame} with dataset containing the variables described in
the model formula}

\item{errorFamily}{A \code{family} object defining the error distribution of the
regression model and a possible use of a link function.  This parameter can also be
a list with two elements: link and family. Each of these are a \code{character} scalar
assigning the link function and error family respectively.
\code{\link[errorFamilies]{errorFamilies()}} will return a list of error distributions
supported by \code{glmNIMBLE} and the potential link functions associated with them.}

\item{regCoeffs}{A character scalar containing the text \code{"none"},
\code{"ridge"}, or \code{"lasso"}.  This determines the type of regularisation to use for the
regression coefficients in the model.  \code{"none"} (the default) results in wide normal priors
for the regression coefficients.  \code{"ridge"} results in normal priors for each of the
regression coefficients regulated by a shared precision parameter which itself has a wide
gamma hyperprior.  \code{"lasso"} results in Laplace priors for each of the regression
coefficients regulated by a shared rate parameter which itself has a wide gamma hyperprior}
}
\description{
A simple interface to run a generalised linear model using the NIMBLE
implementation of MCMC for Bayesian analysis using standard model notation familiar
to users of the \code{\link[base::glm]{glm}} function.
}
