% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleExtensions.R
\name{parameterTest_vectorDouble}
\alias{parameterTest_vectorDouble}
\title{Test a Vector Input Parameter (Double)}
\usage{
parameterTest_vectorDouble(inValue, valRange, lengthRange, lengthHandle = 0)
}
\arguments{
\item{inValue}{A numeric vector to test the values of.}

\item{valRange}{A numeric vector giving the range of values that the input
parameter can take.  If it is of length one then the value is tested whether
it is larger than \code{valRange}.  If it is of length two then the values of
\code{inValue} is tested between the values of \code{valRange}.}

\item{lengthRange}{An integer vector giving the range of sizes the input
vector can take.  If it is of length one, then the vector must be of the
length given.  If it is of length two, then the vector length must be within
the given values.}

\item{lengthHandle}{A scalar integer.  If the value is 0 an error is given
if the length requirements are not met.  If the value is 1 then the values
are recycled to the correct length.  If the value is 2 then the vector is
padded with zeros to the correct length.}
}
\value{
The input value (recycled and padded if appropriate).
}
\description{
Internal function to test whether an input vector parameter
has valid values and dimension.
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
\keyword{internal}
