% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleExtensions.R
\name{rbetabin}
\alias{rbetabin}
\title{Probability Density of the Beta-Binomial Distribution ----}
\usage{
rbetabin(n, shape1, shape2, size)
}
\arguments{
\item{n}{A scalar integer containing the number of variables to simulate (NIMBLE
currently only allows n = 1)}

\item{shape1}{A scalar value containing the first shape parameter of the composite
beta distribution}

\item{shape2}{A scalar value containing the second shape parameter of the composite
beta distribution}

\item{size}{A scalar integer containing the number of trials from the composite
binomial distribution}
}
\value{
A scalar containing the simulated variable
}
\description{
Generate a varibale from the beta-binomial distribution.
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
