% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomialEcosystemState.R
\name{simulateMultinomialEcosystemState}
\alias{simulateMultinomialEcosystemState}
\title{Simulate Instances for the Multinomial Ecosystem State Model}
\usage{
simulateMultinomialEcosystemState(
  numSims,
  stateValModels,
  stateProbModels,
  statePrecModels,
  inputData,
  numStates = NULL,
  stateValError = gaussian,
  coefficientValues = NULL
)
}
\arguments{
\item{numSims}{The number of simulations to draw from the multinomial ecosystem state model.}

\item{stateValModels}{A formula describing the regression relationship between the mean
ecosystem state value and the covariates for all ecosystem states, or a list of formulas
with each element giving the regression relationship between the mean ecosystem state
value and the covariates for each ecosystem state (ordered according to intercept of the
ecosystem state value on the y-axis).  The ecosystem state variable must be given on the
left-hand side of the formula.}

\item{stateProbModels}{A formula describing the regression relationship between the probability
of the existence of ecosystem state and the covariates for all ecosystem states, or a list
of formulas with each element giving the regression relationship between the probability
of the existence of ecosystem state and the covariates for each ecosystem state (ordered
according to intercept of the ecosystem state value on the y-axis).  The ecosystem state
variable must be given on the left-hand side of the formula.}

\item{statePrecModels}{A formula describing the regression relationship between the variance
in the ecosystem state value and the covariates for all ecosystem states, or a list of
formulas with each element giving the regression relationship between the variance in the
ecosystem state value and the covariates for each ecosystem state (ordered according to
intercept of the ecosystem state value on the y-axis).  The ecosystem state variable must
be given on the left-hand side of the formula.}

\item{inputData}{A data frame containing the covariate information and ecosystem state
variable.}

\item{numStates}{An integer scalar containing the number of stable states in the ecosystem.
If any of the \code{stateValModels}, \code{stateProbModels}, or \code{statePrecModels} parameters
is a list then \code{numStates} can be omitted and is therefore set to the length of the
list.}

\item{stateValError}{A description of the error distribution and link function to be used
in the model describing the ecosystem state value.  This can be from the \link[stats]{family}
specification or \code{character} scalar with the following possible values: \code{"gaussian"},
\code{"gamma"}, \code{"beta"}, \code{"negbinomial"}, or \code{"betabinomial"}.}

\item{coefficientValues}{A list containing the values of the coefficients to use in the
simulation.}
}
\value{
A list containg a vector of simulated values for each stochastic node.  In addition the
following elements are appended to the list:
\itemize{
\item{\code{linStateVal}}{A matrix containing the predicted ecosystem state values at each row
of the input covariate data.frame.  Each column represents the predicted ecosystem state value
for each stable state}
\item{\code{linStatePrec}}{A matrix containing the predicted precision of the ecosystem state
value at each row of the input covariate data.frame.  Each column represents the predicted
precision of the ecosystem state value for each stable state}
\item{\code{linStateProb}}{A matrix containing the probability of each ecosystem state existing
at each row of the input covariate data.frame.  Each column represents the probability of each
ecosystem state existing for each stable state}
}
}
\description{
This function takes a set of model specifications for the three sub-model
components of the multinomial ecosystem state model and generates a simulation from this
model specification
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
